% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomisation.R
\name{mass_weighted_urn_design}
\alias{mass_weighted_urn_design}
\title{Generate allocation sequence according to
mass weighted urn design (MWU)}
\usage{
mass_weighted_urn_design(target_alloc, sample_size, alpha = 4)
}
\arguments{
\item{target_alloc}{The target allocation ratio vector}

\item{sample_size}{The total number of allocations to produce}

\item{alpha}{The level of imbalance control recommend at least 4.}
}
\value{
A list containing:
\item{max_imbalance_bound}{Theoretical imbalance bound}
\item{imbalance}{Calculated imbalance}
\item{alloc_predict}{Euclidean allocation predictability}
\item{rand_num}{The random number generated}
\item{trt}{Vector of treatment allocations}
\item{mass}{Matrix of urn masses at each iteration}
\item{sample_size}{The sample size in each arm at each iteration}
\item{selection_prob}{Conditional selection probability at next iteration}
}
\description{
Generate allocation sequence according to
mass weighted urn design (MWU)
}
