% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_probs.R
\name{prob_all_equivalent}
\alias{prob_all_equivalent}
\title{Probability that all cells are equivalent}
\usage{
prob_all_equivalent(mat, eps)
}
\arguments{
\item{mat}{A matrix of cell mean posterior draws.}

\item{eps}{The equivalence margin.}
}
\value{
The probability that all cells in \code{mat}
  are equivalent within the bound \code{eps}.
}
\description{
Probability that all cells are equivalent
}
