% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_probs.R
\name{prob_all_noninferior}
\alias{prob_all_noninferior}
\title{Probability that all cells are noninferior.}
\usage{
prob_all_noninferior(mat, col, eps)
}
\arguments{
\item{mat}{A matrix of cell mean posterior draws.}

\item{col}{The reference column for noninferiority.}

\item{eps}{The noninferiority bound.}
}
\value{
The probability that all cells in \code{mat}
  are noninferior to the reference cell \code{col} by
  a non-inferiority margin of \code{eps}.
}
\description{
Probability that all cells are noninferior.
}
