% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_probs.R
\name{prob_best}
\alias{prob_best}
\title{Probability that each cell is best}
\usage{
prob_best(mat)
}
\arguments{
\item{mat}{A matrix of cell mean posterior draws.}
}
\value{
The probability that each column of \code{mat}
  is the maximum.
}
\description{
Probability that each cell is best
}
