% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_probs.R
\name{prob_superior}
\alias{prob_superior}
\title{Probability that one cell is superior}
\usage{
prob_superior(mat, col, eps)
}
\arguments{
\item{mat}{A matrix of cell mean posterior draws.}

\item{col}{The columnm being assessed for superiority.}

\item{eps}{The superiority margin.}
}
\value{
The probability that \code{col} is superior to
  to the other columns in \code{mat} by margin \code{eps}.
}
\description{
Probability that one cell is superior
}
